backend-flask/requirements.txt
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.5
SQLAlchemy==2.0.20
alembic==1.11.1
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
PyJWT==2.8.0
email-validator==1.3.1
itsdangerous==2.1.2
